refactor: :recycle: migrate About Page to typescript, optimize images to webp
